+++ ci4/app/Entities/User.php
@@ -13,9 +13,13 @@ use App\Models\UserModel;
  * @property string $scope
  * @property string $type
  * @property bool $renew_password
+ * @property string $mfa_secret_hash
  *
  *  Many
  * @property RbacRole $rbac_roles
+ *
+ * OTF
+ * @property bool $has_mfa_secret_hash
  */
 class User extends \RestExtension\Entities\User {
 
@@ -69,7 +73,7 @@ class User extends \RestExtension\Entities\User {
         ];
     }
 
-    public $hiddenFields = ['password'];
+    public $hiddenFields = ['password', 'mfa_secret_hash'];
 
     public function getScopes(): array {
         return strlen($this->scope) ? explode(' ', $this->scope) : [];
@@ -95,6 +99,66 @@ class User extends \RestExtension\Entities\User {
             $this->first_name,
             $this->username
         );
+    }
+
+    public function hasMFASecret(): bool {
+        return strlen($this->mfa_secret_hash) > 0;
+    }
+
+    public function updateMFASecret(string $value): void {
+        $cipher = 'AES-256-CBC';
+        $passPhrase = hex2bin('c17319112b52d6b472136d7938121233783d723814746a67c81a451c4d238d18');
+        $nonceSize = openssl_cipher_iv_length($cipher);
+        $nonce = openssl_random_pseudo_bytes($nonceSize);
+
+        $ciphertext = openssl_encrypt(
+            $value,
+            $cipher,
+            $passPhrase,
+            OPENSSL_RAW_DATA,
+            $nonce
+        );
+
+        $this->mfa_secret_hash = base64_encode($nonce . $ciphertext);
+        $this->save();
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function getMFASSecret(): string {
+        $cipher = 'AES-256-CBC';
+        $passPhrase = hex2bin('c17319112b52d6b472136d7938121233783d723814746a67c81a451c4d238d18');
+
+        $message = base64_decode($this->mfa_secret_hash, true);
+        if ($message === false) {
+            throw new \Exception('Encryption failure');
+        }
+
+        $nonceSize = openssl_cipher_iv_length($cipher);
+        $nonce = mb_substr($message, 0, $nonceSize, '8bit');
+        $ciphertext = mb_substr($message, $nonceSize, null, '8bit');
+
+        $plaintext = openssl_decrypt(
+            $ciphertext,
+            $cipher,
+            $passPhrase,
+            OPENSSL_RAW_DATA,
+            $nonce
+        );
+
+        return $plaintext;
+    }
+
+    public function removeMFASecret(): void {
+        $this->mfa_secret_hash = null;
+        $this->save();
+    }
+
+    public function toArray(bool $onlyChanged = false, bool $cast = true, bool $recursive = false, array $fieldsFilter = null): array {
+        $item = parent::toArray($onlyChanged, $cast, $recursive, $fieldsFilter);
+        $item['has_mfa_secret_hash'] = $this->hasMFASecret();
+        return $item;
     }
 
     /**