+++ ci4/app/Interfaces/UsersMFASetupPrepareResponse.php
@@ -1,0 +1,12 @@
+<?php namespace App\Interfaces;
+
+/**
+ * Interface UsersMFASetupPrepareResponse
+ * @package App\Interfaces
+ * @property bool $hasMFA
+ * @property ?string $qrCodeDataUri
+ * @property ?string $setupCode
+ */
+interface UsersMFASetupPrepareResponse {
+
+}