+++ ci4/app/Views/Login/Login.php
@@ -290,26 +290,6 @@ use App\Libraries\EmailLib;
             right: 24px;
         }
 
-        .logo-header {
-            font-size: 20px;
-            color: #ffffff;
-            float: left;
-            text-transform: uppercase;
-            content: ' ';
-            background: url(/admin_assets/img/logo-header-blue.svg);
-
-            background-repeat: no-repeat;
-            background-size: auto 100%;
-            /* background-position: 103px 0px; */
-            width: 100%;
-            height: 60px;
-            display: block;
-            opacity: 1;
-            margin: 0 !important;
-            background-position: center center;
-            background-size: contain;
-        }
-
         .btn-secondary a, .btn-secondary a:hover {
             color: white;
         }
@@ -330,7 +310,7 @@ use App\Libraries\EmailLib;
                 } else {
                     field.type = "password";
                 }
-            })
+            });
         });
     </script>
 </head>