+++ vue/src/core/services/Deploy/models/DeploymentsLabel.ts
@@ -1,0 +1,14 @@
+/**
+ * Created by ModelParser
+ * Date: 12-10-2024.
+ * Time: 10:37.
+ */
+import {DeploymentsLabelDefinition} from './definitions/DeploymentsLabelDefinition';
+
+export class DeploymentsLabel extends DeploymentsLabelDefinition {
+
+    constructor(json?: any) {
+        super(json);
+    }
+
+}